+++ hata/discord/emoji/unicode/tests/test__Unicode__constructor.py
@@ -3,10 +3,22 @@ import vampytest
 from ..unicode_type import Unicode
 
 
-def test__Unicode__repr():
+def test__Unicode__new():
     """
-    Tests whether ``Unicode``'ss constructor works as intended.
+    Tests whether ``Unicode``'s `__repr__` works as intended.
     """
-    unicode = Unicode('owo', b'56', False, ('anya',), ('got', 'banned'))
+    name = 'owo'
+    value = '56'
+    variation_selector_16 = False
+    emoticons = ('anya',)
+    aliases = ('got', 'banned')
+    unicode_aliases = ('54', '53')
     
-    vampytest.assert_instance(repr(unicode), str)
+    unicode = Unicode(name, value, variation_selector_16, aliases, emoticons, unicode_aliases)
+    
+    vampytest.assert_eq(unicode.name, name)
+    vampytest.assert_eq(unicode.value, value)
+    vampytest.assert_eq(unicode.variation_selector_16, variation_selector_16)
+    vampytest.assert_eq(unicode.emoticons, emoticons)
+    vampytest.assert_eq(unicode.aliases, aliases)
+    vampytest.assert_eq(unicode.unicode_aliases, unicode_aliases)