+++ hata/discord/emoji/unicode/tests/test__Unicode__magic.py
@@ -3,21 +3,10 @@ import vampytest
 from ..unicode_type import Unicode
 
 
-def test__Unicode__new__0():
+def test__Unicode__repr():
     """
-    Tests whether ``Unicode``'s `__repr__` works as intended.
+    Tests whether ``Unicode``'ss constructor works as intended.
     """
-    name = 'owo'
-    value_raw = b'56'
-    value = '56'
-    variation_selector_16 = False
-    emoticons = ('anya',)
-    aliases = ('got', 'banned')
+    unicode = Unicode('owo', '56', False, ('anya',), ('got', 'banned'), ('54', '53'))
     
-    unicode = Unicode(name, value_raw, variation_selector_16, aliases, emoticons)
-    
-    vampytest.assert_eq(unicode.name, name)
-    vampytest.assert_eq(unicode.value, value)
-    vampytest.assert_eq(unicode.variation_selector_16, variation_selector_16)
-    vampytest.assert_eq(unicode.emoticons, emoticons)
-    vampytest.assert_eq(unicode.aliases, aliases)
+    vampytest.assert_instance(repr(unicode), str)