+++ hata/discord/emoji/unicode/tests/test__Unicode__utility.py
@@ -8,7 +8,7 @@ def test__Unicode__get_system_name():
     Tests whether ``Unicode``'s `.get_system_name` works as intended.
     """
     name = 'owo'
-    unicode = Unicode(name, b'', False, None, None)
+    unicode = Unicode(name, '', False, None, None, None)
     
     system_name = unicode.get_system_name()
     vampytest.assert_instance(system_name, str)
@@ -21,7 +21,7 @@ def test__Unicode__iter_emoticons__0():
     
     Case: 0
     """
-    unicode = Unicode('', b'', False, None, None)
+    unicode = Unicode('', '', False, None, None, None)
     
     vampytest.assert_eq({*unicode.iter_emoticons()}, set())
 
@@ -33,7 +33,7 @@ def test__Unicode__iter_emoticons__1():
     Case: 1+
     """
     emoticons = ('owo', 'uwu')
-    unicode = Unicode('', b'', False, None, emoticons)
+    unicode = Unicode('', '', False, None, emoticons, None)
     
     vampytest.assert_eq({*unicode.iter_emoticons()}, {*emoticons})
 
@@ -44,7 +44,7 @@ def test__Unicode__iter_aliases__0():
     
     Case: 0
     """
-    unicode = Unicode('', b'', False, None, None)
+    unicode = Unicode('', '', False, None, None, None)
     
     vampytest.assert_eq({*unicode.iter_aliases()}, set())
 
@@ -56,7 +56,7 @@ def test__Unicode__iter_aliases__1():
     Case: 1+
     """
     aliases = ('owo', 'uwu')
-    unicode = Unicode('', b'', False, aliases, None)
+    unicode = Unicode('', '', False, aliases, None, None)
     
     vampytest.assert_eq({*unicode.iter_aliases()}, {*aliases})
 
@@ -67,7 +67,7 @@ def test__Unicode__iter_alternative_names__0():
     
     Case: 0
     """
-    unicode = Unicode('', b'', False, None, None)
+    unicode = Unicode('', '', False, None, None, None)
     
     vampytest.assert_eq({*unicode.iter_alternative_names()}, set())
 
@@ -81,6 +81,31 @@ def test__Unicode__iter_alternative_names__1():
     emoticons = ('owo', 'uwu')
     aliases = ('awa', 'iwi')
     
-    unicode = Unicode('', b'', False, aliases, emoticons)
+    unicode = Unicode('', '', False, aliases, emoticons, None)
     
     vampytest.assert_eq({*unicode.iter_alternative_names()}, {*emoticons, *aliases})
+
+
+
+def test__Unicode__iter_unicode_aliases__0():
+    """
+    Tests whether ``Unicode``'s `.iter_unicode_aliases` works as intended.
+    
+    Case: 0
+    """
+    unicode = Unicode('', '', False, None, None, None)
+    
+    vampytest.assert_eq({*unicode.iter_unicode_aliases()}, set())
+
+
+def test__Unicode__iter_unicode_aliases__1():
+    """
+    Tests whether ``Unicode``'s `.iter_unicode_aliases` works as intended.
+    
+    Case: 1+
+    """
+    unicode_alternatives = ('owo', 'uwu')
+    
+    unicode = Unicode('', '', False, None, None, unicode_alternatives)
+    
+    vampytest.assert_eq({*unicode.iter_unicode_aliases()}, {*unicode_alternatives})